+++ framework/src/mesh/degree_of_freedom.cc
@@ -6,7 +6,7 @@ namespace ffea {
 
 DegreeOfFreedom::DegreeOfFreedom() {}
 
-DegreeOfFreedom::DegreeOfFreedom(int local_id, short number_of_auxiliary_values)
+DegreeOfFreedom::DegreeOfFreedom(size_t local_id, size_t number_of_auxiliary_values)
     : value_(),
       auxiliary_values_(number_of_auxiliary_values),
       local_id_(local_id),
@@ -15,9 +15,9 @@ DegreeOfFreedom::DegreeOfFreedom(int local_id, short number_of_auxiliary_values)
 
 DegreeOfFreedom::~DegreeOfFreedom() {}
 
-int DegreeOfFreedom::local_id() const { return local_id_; }
+size_t DegreeOfFreedom::local_id() const { return local_id_; }
 
-int DegreeOfFreedom::global_id() const { return global_id_; }
+size_t DegreeOfFreedom::global_id() const { return global_id_; }
 
 double DegreeOfFreedom::value() const { return value_; }
 
@@ -27,19 +27,19 @@ void DegreeOfFreedom::set_value(const Eigen::VectorXd &solution) {
   set_value(solution(local_id()));
 }
 
-double DegreeOfFreedom::auxiliary_value(short index) const {
+double DegreeOfFreedom::auxiliary_value(size_t index) const {
   check_auxiliary_value(index);
 
   return auxiliary_values_[index];
 }
 
-void DegreeOfFreedom::set_auxiliary_value(short index, double value) {
+void DegreeOfFreedom::set_auxiliary_value(size_t index, double value) {
   check_auxiliary_value(index);
 
   auxiliary_values_[index] = value;
 }
 
-void DegreeOfFreedom::check_auxiliary_value(short index) const {
+void DegreeOfFreedom::check_auxiliary_value(size_t index) const {
   if (auxiliary_values_.empty()) {
     throw std::out_of_range("There are no auxiliary values.");
   }
